chore: rename RisAutoComplete, remove fluid prop
