--- src/components/AutoComplete/index.ts
@@ -1,3 +1,0 @@
-import AutoComplete from "./AutoComplete.vue";
-
-export default AutoComplete;
