+++ src/components/RisAutoComplete/RisAutoComplete.unit.spec.ts
@@ -2,7 +2,7 @@ import { describe, it, expect } from "vitest";
 import { mount, shallowMount } from "@vue/test-utils";
 
 import { nextTick } from "vue";
-import AutoComplete, { type Props } from "./AutoComplete.vue";
+import AutoComplete, { type Props } from "./RisAutoComplete.vue";
 import PrimeVue from "primevue/config";
 import ProgressSpinner from "primevue/progressspinner";
 import _ from "lodash";
@@ -61,7 +61,6 @@ describe("AutoComplete", () => {
   it("passes props correctly", async () => {
     const props: Required<Props> = {
       suggestions: [{ id: "1", label: "Option 1" }],
-      fluid: false,
       typeahead: false,
       dropdownMode: "current",
       ariaLabel: "ARIA label",
