+++ src/components/RisAutoComplete/index.ts
@@ -1,0 +1,3 @@
+import RisAutoComplete from "./RisAutoComplete.vue";
+
+export default RisAutoComplete;
